[Clang] Unify 'nvptx-arch' and 'amdgpu-arch' into 'offload-arch'
Summary:
These two tools do the same thing, we should unify them into a single
tool. We create symlinks for backward compatiblity and provide a way to
get the old vendor specific behavior with `--amdgpu-only` and
`--nvptx-only`.
