+++ clang/tools/offload-arch/AMDGPUArchByHIP.cpp
