+++ clang/tools/offload-arch/AMDGPUArchByKFD.cpp
